+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added basic support for `TypedArray`s and `Set` values in `struct`, `signature` and `table` views
 - Fixed displaying supported syntaxes in view's showcase for `source` view
+- Exposed used CodeMirror for extension purposes via a static property `CodeMirror` of `QueryEditor` and `ViewEditor`, e.g. `discovery.view.QueryEditor.CodeMirror.defineMode('my-mode', ...)`
 
 ## 1.0.0-beta.80 (21-01-2024)