+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Modified the badge view to enable passing all options via config, in addition to data
 - Fixed the highlighting of Jora assertions
 - Fixed displaying supported syntaxes in view's showcase for `source` view
+- Fixed passing context to `table` view when rendered in `struct` view
 - Exposed used CodeMirror for extension purposes via a static property `CodeMirror` of `QueryEditor` and `ViewEditor`, e.g. `discovery.view.QueryEditor.CodeMirror.defineMode('my-mode', ...)`
 
 ## 1.0.0-beta.80 (21-01-2024)