+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added basic support for `TypedArray`s and `Set` values in `struct`, `signature` and `table` views
 - Added `'only'` value for `darkmode` option of `Widget` & `App`, which forces to use dark mode with no option for a user to switch to light mode
+- Added `scalarCol` option for `table` view to display the row value as the first cell
 - Modified the badge view to enable passing all options via config, in addition to data
 - Fixed the highlighting of Jora assertions
 - Fixed displaying supported syntaxes in view's showcase for `source` view